refactor: add usePianoNotesKeyboardShortcut
